+++ sdk/tlkapi/tlkapi_type.h
@@ -70,13 +70,13 @@ typedef unsigned long        ulong;
 
 typedef signed char          sint08;
 typedef signed short         sint16;
-typedef signed int           sint32;
+typedef signed long          sint32;
 typedef signed long long     sint64;
 
 
 typedef unsigned char		 uint08;
 typedef unsigned short		 uint16;
-typedef unsigned int		 uint32;
+typedef unsigned long		 uint32;
 typedef unsigned long long	 uint64;
